+++ __vp/includes/error.php
@@ -13,13 +13,13 @@ header('Status: 404 Not Found');
         <title>Error</title>
         <style type="text/css">
             body{
-                background: #eee;
-                font-family: Helvetica;
+                background: #f5f5f5;
+                font-family: Tahoma, Helvetica, Arial, sans-serif;
             }
             .main{
                 background: #fff;
-                box-shadow: 0 0 25px #aaa;
-                border: 1px solid #888;
+                box-shadow: 0 0 25px #bbb;
+                border: 1px solid #aaa;
                 width: 40%;
                 border-radius: 5px;
                 margin: 10% auto;
@@ -30,9 +30,9 @@ header('Status: 404 Not Found');
     </head>
     <body>
         <div class="main">
-            <h1>Page not found</h1> 
-            <h1>OR</h1>
-            <h1>Down for Maintain</h1>
+            <h1>Page Not Found</h1> 
+            <h1>or</h1>
+            <h1>Down For Maintain</h1>
         </div>
     </body>
 </html>