+++ _apps/default/view/error.php
@@ -9,9 +9,7 @@ header('Status: 404 Not Found');
 $this->META_DETAILS['TITLE'] = 'Error';
 $this->META_DETAILS['DESCRIPTION'] = '';
 $this->META_DETAILS['KEYWORDS'] = '';
-
 ?>
-
 <div class="main">
     <h1>404 Error</h1>
     <h1>Page not found</h1>