+++ _apps/default/view/home.php
@@ -5,7 +5,7 @@ if (!defined('ROOT')) {
 ?>
 <div class="main">
 
-    <h2 style="text-align: center">Welcome to Varphp ( $P )</h2>
+    <h2 style="text-align: center">Welcome to Varphp ($P)</h2>
 
     <h4>URL: <?php echo $this->URL('APP'); ?></h4>
 
@@ -13,6 +13,8 @@ if (!defined('ROOT')) {
 
     <h4>PATH: <?php echo PATH; ?></h4>
 
-    <div>Create <code style="font-weight: bold; font-size: 16px;">`_config.php`</code> file and Class in root directory and set Configuration.</div>
+    <div>Create <code style="font-weight: bold; font-size: 16px;">`_config.php`</code> 
+        file and Class in root directory and set Configuration.
+    </div>
 
 </div>