+++ index.php
@@ -1,8 +1,14 @@
 <?php
 
-define('ROOT', str_replace('\\', '/', dirname(__FILE__)) . '/');
+define('ROOT', str_ireplace('\\', '/', dirname(__FILE__)) . '/');
 
-define('PATH', dirname($_SERVER['SCRIPT_NAME']) . '/');
+$SCRIPT_NAME = str_ireplace('\\', '/', dirname($_SERVER['SCRIPT_NAME']));
+if (strlen($SCRIPT_NAME) > 1) {
+    define('PATH', $SCRIPT_NAME . '/');
+} else {
+    define('PATH', $SCRIPT_NAME);
+}
+unset($SCRIPT_NAME);
 
 require_once '__vp/system/boot.php';
 ?>