+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log & History
 
+#### 07-11-2016
+    Stable Version 2.3
+    Few Minor Bugs Fixed
+
+    Improved:
+        `index.php` File Edited
+        index Controller
+
 #### 01-11-2016
     Stable Version 2.2
     Stability Improved
@@ -16,7 +24,7 @@
 
 #### 27-09-2016
 	Beta Version 2.0
-	
+
 #### 2013
 	Founded
 	First made